Move the queue to its own file
